requirements.txt
# Required Python packages for tsetools
Flask
werkzeug.utils
setuptools